Extracted adding functionality to a manager class.
